# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HSE Platform API
  description: "\nHSE platform API — Phase 0 Foundation (spec `docs/specs/0-foundation.md` v1.0).\n\n\
    * Auth: `POST /api/v1/auth/login` sets the httpOnly SameSite=Lax cookie `hse_session` (JWT) and\n\
    \  returns `{access_token, user}`. Send the cookie or `Authorization: Bearer <token>`.\n* Errors:\
    \ `{\"detail\": {\"code\", \"message\", \"message_ar\", \"errors\"}}`; 401 unauthenticated,\n  403\
    \ forbidden, 404 not found / out of scope, 409 conflict / invalid transition, 422 validation.\n* Lists:\
    \ `?page=1&page_size=50` → `{items, total, page, page_size}`.\n* Timestamps are UTC; display and day\
    \ boundaries use the project timezone (Asia/Riyadh).\n"
  version: 0.1.0
paths:
...
  /api/v1/projects/{project_id}/sites:
    get:
      tags:
      - sites & zones
      summary: List sites of a project (restricted to assigned sites when scoped)
      operationId: list_sites
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SiteStatus'
          - type: 'null'
          title: Status
      - name: site_side
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SiteSide'
          - type: 'null'
          title: Site Side
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          description: Code/name search.
          title: Q
        description: Code/name search.
      - name: sort
        in: query
        required: false
        schema:
          enum:
          - code
          - -code
          - name
          - -name
          type: string
          default: code
          title: Sort
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 1-based page number.
          default: 1
          title: Page
        description: 1-based page number.
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Items per page.
          default: 50
          title: Page Size
        description: Items per page.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SitePage'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - sites & zones
      summary: Create a site
      operationId: create_site
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/sites/{site_id}:
    get:
      tags:
      - sites & zones
      summary: Get a site
      operationId: get_site
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Site Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - sites & zones
      summary: Update a site
      operationId: update_site
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Site Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/sites/{site_id}/transitions:
    post:
      tags:
      - sites & zones
      summary: Activate / inactivate a site
      operationId: transition_site
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Site Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteTransitionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}/zones:
    get:
      tags:
      - sites & zones
      summary: List zones of a project (restricted to assigned sites when scoped)
      operationId: list_zones
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: site_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Site Id
      - name: zone_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ZoneType'
          - type: 'null'
          title: Zone Type
      - name: airside_area
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/AirsideArea'
          - type: 'null'
          title: Airside Area
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ZoneStatus'
          - type: 'null'
          title: Status
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          description: Code/name search.
          title: Q
        description: Code/name search.
      - name: sort
        in: query
        required: false
        schema:
          enum:
          - code
          - -code
          - name
          - -name
          type: string
          default: code
          title: Sort
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 1-based page number.
          default: 1
          title: Page
        description: 1-based page number.
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Items per page.
          default: 50
          title: Page Size
        description: Items per page.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZonePage'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/sites/{site_id}/zones:
    post:
      tags:
      - sites & zones
      summary: Create a zone in a site
      operationId: create_zone
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Site Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/zones/{zone_id}:
    get:
      tags:
      - sites & zones
      summary: Get a zone
      operationId: get_zone
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Zone Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - sites & zones
      summary: Update a zone (airside attribute changes audited with before/after)
      operationId: update_zone
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Zone Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/zones/{zone_id}/transitions:
    post:
      tags:
      - sites & zones
      summary: Close temporarily / reopen / archive a zone
      operationId: transition_zone
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Zone Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneTransitionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    AirsideArea:
      type: string
      enum:
      - runway
      - runway_strip
      - resa
      - taxiway
      - taxiway_strip
      - apron
      - ils_critical
      - ils_sensitive
      - airside_road
      - other_airside
      title: AirsideArea
...
    ErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Envelope for every 4xx/5xx response.
...
    SiteCreate:
      properties:
        code:
          type: string
          pattern: ^[A-Za-z0-9-]{1,12}$
          title: Code
          description: Unique within the project, e.g. S-AIR.
        name_en:
          type: string
          maxLength: 150
          minLength: 1
          title: Name En
        name_ar:
          type: string
          maxLength: 150
          minLength: 1
          title: Name Ar
        site_side:
          $ref: '#/components/schemas/SiteSide'
          description: airside/mixed only on airport projects.
        gps_lat:
          anyOf:
          - type: number
            maximum: 33.0
            minimum: 16.0
          - type: 'null'
          title: Gps Lat
          description: KSA bbox latitude.
        gps_lng:
          anyOf:
          - type: number
            maximum: 56.0
            minimum: 34.0
          - type: 'null'
          title: Gps Lng
          description: KSA bbox longitude.
      additionalProperties: false
      type: object
      required:
      - code
      - name_en
      - name_ar
      - site_side
      title: SiteCreate
    SitePage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SiteRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
          description: Total number of matching items across all pages.
        page:
          type: integer
          title: Page
          description: 1-based page number.
        page_size:
          type: integer
          title: Page Size
          description: Items per page (max 200).
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: SitePage
    SiteRead:
      properties:
        created_at:
          type: string
          format: date-time
          title: Created At
          description: UTC timestamp.
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: UTC timestamp.
        id:
          type: string
          format: uuid
          title: Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        code:
          type: string
          title: Code
        name_en:
          type: string
          title: Name En
        name_ar:
          type: string
          title: Name Ar
        site_side:
          $ref: '#/components/schemas/SiteSide'
        gps_lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Gps Lat
        gps_lng:
          anyOf:
          - type: number
          - type: 'null'
          title: Gps Lng
        status:
          $ref: '#/components/schemas/SiteStatus'
      type: object
      required:
      - created_at
      - updated_at
      - id
      - project_id
      - code
      - name_en
      - name_ar
      - site_side
      - gps_lat
      - gps_lng
      - status
      title: SiteRead
    SiteSide:
      type: string
      enum:
      - airside
      - landside
      - mixed
      - other
      title: SiteSide
    SiteStatus:
      type: string
      enum:
      - active
      - inactive
      title: SiteStatus
    SiteTransitionRequest:
      properties:
        to_status:
          $ref: '#/components/schemas/SiteStatus'
        reason:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Reason
      additionalProperties: false
      type: object
      required:
      - to_status
      title: SiteTransitionRequest
      description: active ↔ inactive. Sites are never deleted (rule 23).
    SiteUpdate:
      properties:
        code:
          anyOf:
          - type: string
            pattern: ^[A-Za-z0-9-]{1,12}$
          - type: 'null'
          title: Code
        name_en:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Name En
        name_ar:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Name Ar
        site_side:
          anyOf:
          - $ref: '#/components/schemas/SiteSide'
          - type: 'null'
        gps_lat:
          anyOf:
          - type: number
            maximum: 33.0
            minimum: 16.0
          - type: 'null'
          title: Gps Lat
        gps_lng:
          anyOf:
          - type: number
            maximum: 56.0
            minimum: 34.0
          - type: 'null'
          title: Gps Lng
      additionalProperties: false
      type: object
      title: SiteUpdate
...
    ZoneCreate:
      properties:
        code:
          type: string
          pattern: ^[A-Za-z0-9-]{1,16}$
          title: Code
          description: Unique within the site, e.g. Z-APR-21.
        name_en:
          type: string
          maxLength: 150
          minLength: 1
          title: Name En
        name_ar:
          type: string
          maxLength: 150
          minLength: 1
          title: Name Ar
        zone_type:
          $ref: '#/components/schemas/ZoneType'
          description: airside/landside only on airport projects and consistent with site_side.
        airside:
          anyOf:
          - $ref: '#/components/schemas/AirsideAttributesInput'
          - type: 'null'
          description: Required iff zone_type = airside; must be null otherwise.
      additionalProperties: false
      type: object
      required:
      - code
      - name_en
      - name_ar
      - zone_type
      title: ZoneCreate
    ZonePage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ZoneRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
          description: Total number of matching items across all pages.
        page:
          type: integer
          title: Page
          description: 1-based page number.
        page_size:
          type: integer
          title: Page Size
          description: Items per page (max 200).
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: ZonePage
    ZoneRead:
      properties:
        created_at:
          type: string
          format: date-time
          title: Created At
          description: UTC timestamp.
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: UTC timestamp.
        id:
          type: string
          format: uuid
          title: Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        site_id:
          type: string
          format: uuid
          title: Site Id
        code:
          type: string
          title: Code
        name_en:
          type: string
          title: Name En
        name_ar:
          type: string
          title: Name Ar
        zone_type:
          $ref: '#/components/schemas/ZoneType'
        status:
          $ref: '#/components/schemas/ZoneStatus'
        status_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Reason
        airside:
          anyOf:
          - $ref: '#/components/schemas/AirsideAttributes'
          - type: 'null'
          description: Present iff zone_type = airside.
      type: object
      required:
      - created_at
      - updated_at
      - id
      - project_id
      - site_id
      - code
      - name_en
      - name_ar
      - zone_type
      - status
      - status_reason
      - airside
      title: ZoneRead
    ZoneStatus:
      type: string
      enum:
      - active
      - temporarily_closed
      - archived
      title: ZoneStatus
      description: Zone lifecycle (§4.4).
    ZoneTransitionRequest:
      properties:
        to_status:
          $ref: '#/components/schemas/ZoneStatus'
        reason:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Reason
      additionalProperties: false
      type: object
      required:
      - to_status
      title: ZoneTransitionRequest
      description: active → temporarily_closed → active; active/temporarily_closed → archived (§4.4).
    ZoneType:
      type: string
      enum:
      - airside
      - landside
      - other
      title: ZoneType
      description: airside/landside only on airport projects (§5.3 rule 18).
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: httpOnly SameSite=Lax cookie set by POST /auth/login.
      in: cookie
      name: hse_session
    bearerAuth:
      type: http
      description: 'Same JWT as the cookie, as `Authorization: Bearer <token>`.'
      scheme: bearer